Show proportion of jobs cancelled by dependency
- Add functionality for the app to read the CSV file of jobs.

- Add calculation to aggregate the above into a single-column table of
job requests, the column being the proportion of jobs cancelled by
dependency over the total number of jobs that either errored or were
cancelled by dependency. Job requests with no jobs that errored or
cancelled by dependency are not of interest and hence excluded.

- Show the distribution of the proportion of jobs that were cancelled
by dependency in a histogram in the Streamlit app.