Test `tasks.get_project_definitions.get_query`
This adds adds a test and removes the pragma. To do so, it also adds two
fixtures that provide access to the testing jobserver DB (014b30c) and
to the objects representing jobserver's tables (07eaf3f).

We have to take a slightly different approach to make the testing
jobserver DB available in CI. Rather than configuring it in
`docker/docker-compose.yaml`, we configure it in
`.github/workflows/ci.yaml`. The result is some tight coupling between
these files and `tests/conftest.py`. Given that we don't expect these
files to change much, we think that this tight coupling is okay.