Place `app` directory on `PYTHONPATH`
The `streamlit run` command expects imported modules to be at the same
level as the Streamlit app module. pytest, however, doesn't. It's easier
to configure pytest than Streamlit, so we do so by placing the `app`
directory on the `PYTHONPATH` and ignoring the slightly confusing
imports in the Streamlit app module.