Make `test_extract` more specific
Previously, the test passed whether or not *the correct* row was
returned by the query. All that mattered was that *a* row was returned
by the query. We make the test more specific by assigning different SHAs
to different job requests. We also emphasise that whilst the attributes,
and their order, is important, their values aren't (with the exception
of the `sha` attribute).

Co-authored-by: Alice Wong <146962159+alarthast@users.noreply.github.com>