Test `tasks.get_job_requests.extract`
This adds a test and removes the pragma. The test has a substantial
arrange step by necessity and design. By necessity, because there's a
lot to arrange. By design, because we'd like to keep the fixtures
separate from other tests, even though several are very similar to the
fixtures in other tests. Doing so makes it much easier to update this
test, without affecting other tests.

Co-authored-by: Alice Wong <146962159+alarthast@users.noreply.github.com>