Stop trying to open duplicate PRs
To fix a vulnerability, dependabot security updates will first push a
commit to a non-main branch and then open a PR.

Previously, the CI workflow would also try to open a PR in response,
resulting in a race. Now, the CI workflow won't: the `open_pr` job is
skipped when the username of the user that triggered the initial
workflow run is "dependabot".

Of course, it's really hard to test this change without rerunning a
failed job. But even then, I'm not clear which branch the failed job
will be run on. So, I plan to wait-out and observe what happens the next
time Ethelred encounters dependabot security updates.

[1]: https://github.com/opensafely-core/ethelred/actions/runs/16173754737