Refactor `get_stage`
The primary refactoring is to reduce the line count by inlining both a
variable and a condition. The secondary refactoring is to unpack the
first substring returned by `split` into a named variable, which makes
the function body easier to reason about. (It may be ever so slightly
more performant, because we pass `maxsplit`, but it's hard to say for
sure.)