Set `tool.ruff.src`
In Streamlit, first-party imports are rooted at the level of the app
module. (We avoid the terms "relative" and "absolute" because they have
Python-specific meanings.) For example, we use `import charts` rather
than `from app import charts` in `app.app` (the app module) to import
the first-party module `charts`. Ruff doesn't know this: by default, it
expects first-party imports to be rooted in the project root; if they
aren't, then it treats them as third-party imports.

By setting [`tool.ruff.src`][1], ruff will know that imports rooted in
the project root *or* imports rooted at the level of the app module are
first-party imports.

Ultimately, doing so ensures imports are placed in the correct section;
standard library, third-party, or first-party (see
[`tool.ruff.lint.isort`][2]).

[1]: https://docs.astral.sh/ruff/settings/#src
[2]: https://docs.astral.sh/ruff/settings/#lint_isort_section-order