Handle empty `run_command`s when determining stage
Sometimes jobs have empty `run_command`s in the real data, so we should
not assume an action name is always present in the run_command.

If a job has no `run_command`, we can't easily classify it as a
database or analysis job.
We can either define the stage as 'unknown' or leave the
stage field blank - I've opted for the latter for now.