`.from_file` -> `.from_function`
We want to test the Streamlit app against a fake repository. To do so,
we have to switch a real repository for a fake repository somehow.
Because of the way Streamlit apps are run, we can't monkeypatch; and a
"is in testing" switch in our Streamlit app is just too much to bear.

Dependency injection would be a good choice: pass a repository (either
real or fake) to `main`. However, that doesn't work with `.from_file`.
It does work with `.from_function`, with the caveat that we have to move
the imports inside `main` (again, because of the way Streamlit apps are
run). So, that's what we begin to do! A follow up commit will add
dependency injection and switch a real repository for a fake repository.