Test fields in result against fields in `Row`
For each `extract` function, we test the fields in the result object (an
SQLAlchemy object) against the fields in the `Row`, rather than against
an ordinary tuple. This has the advantage of ensuring that the result
objects are consistent with the `Row`s, which makes the other tests that
use `Row`s less brittle.