Extract function
We extract the logic for computing the proportion of actions that were
run as jobs (`num_jobs_over_num_actions`) into the
`get_num_jobs_over_num_actions` function. We do this such that we can
make a couple of assertions about the number of jobs and the number of
actions, and more easily test these assertions.