Extract `id` in `tasks.get_job_requests`
We will use the `id` of job requests to dynamically filter the jobs data
presented in the Streamlit app. The app should expect the `id` column
to be present in DataFrame returned by a repository's `get_job_requests`
method.

Adapt the `get_job_requests` task to meet the app's expectations.

Part of #27