Filter visualisations by job request creation date
Add the ability for a user to select a date range at the top of the
Streamlit app to filter the visualisations by the job request
creation date, as this is a feature in Grafana that we want to
demonstrate the capability of replicating.

Ensure all date formats being compared in the filtering process are
in UTC and in a format that pandas can handle.

Closes #27