Remove `AbstractRepository`
It turns out that the costs outweigh the benefits of
`AbstractRepository`, an abstract base class (ABC). This is because
fakes must implement the abstract methods, but if those abstract methods
aren't called by tests, then coverage fails. A test that exists simply
to exercise a fake is clearly a low-value test.

Let's remove `AbstractRepository` and rely on tried-and-tested duck
typing. Quack quack!