Replace `Repository.calculate_proportions`
In b08ef09 we pushed down a calculation; here, we pull up a calculation.
Why? Calculating a proportion (or a *measure*; a numerator over a
denominator) has very little essential complexity. Whilst we probably
shouldn't do it in Altair, we could (see "[Data Transformations][1]" in
the Altair documentation). We have done so here for expediency, and
anticipate refactoring to do so in Pandas in a follow up commit.

What should be clear is that calculating a proportion in Altair or
Pandas is less complex than doing so in a task. In other words, doing so
in a task has more accidental complexity.

For more in essential complexity and accidental complexity, see "[No
Silver Bullet][2]".

[1]: https://altair-viz.github.io/user_guide/transform/index.html
[2]: https://en.wikipedia.org/wiki/No_Silver_Bullet