Remove type of measurement
Avoid passing the type of measurement for the encoded field. If the type
of measurement is passed when the encoded field doesn't exist in the
data frame, then the end-to-end test will pass when it should fail.

`app/charts.py` now contains an interactive Python session that can be
executed with the doctest module. pytest won't run it by default,
which is okay for now. To run it, use:

```
pytest --doctest-modules app/charts.py
```