Pull up calculation
Here, we pull up the calculation of the number of jobs over the number
of actions, a proportion, to Altair. As with 117cec8, calculating a
proportion has very little essential complexity. Pushing the calculation
down to a task increases accidental complexity; pulling the calculation
up to Altair decreases accidental complexity. So, we pull up the
calculation.