Ensure test repository exhibits data integrity
We need to ensure that the data in a test repository looks like the data
in a production repository. For example, if a job has an FK to a job
request, then a corresponding job request should exist (i.e. the test
repository exhibits data integrity). We also need to avoid flaky tests,
however, so to keep track of FKs and PKs, we assign single-digit PKs to
job requests and double digit PKs to jobs.