Exclude all `abc.abstractmethod`s from coverage
Abstract methods are always overridden - otherwise the subclass cannot
be instantiated (and we have a test for this). Therefore, they do not
need to be included in the coverage report.

I've retained the `NotImplementedError`s, as these will get raised if
a subclass tries to cheat by implementing a method by calling `super()`,
which should not be done.