Replace `outcome` with `status_message_type`
`outcome` flattened the `status`/`status_message` hierarchy, which
turned out to be undesirable: sometimes we want to know how many jobs
failed in total; sometimes how many jobs failed for a reason. To answer
the former question, we don't want to sum the answers to the latter
question for the set of reasons. Doing so would be error-prone.

As `status_message` doesn't contain a classification of status messages,
we classify them within `get_status_message_type`. Admittedly, the
classification is simple. However, `get_status_message_type` gives us a
location (and recommends a mechanism, `match`/`case`) for extending it.