Add functions for `get_workflow_runs` task
Add functions to retrieve workflow runs for GitHub Actions from the
GitHub API. These functions will be used in a yet-to-be-written ETL
pipeline in the `get_workflow_runs` task.

Note that timestamps are written as integers instead of human-readable
strings as a first step to avoid having to parse dates for now.

Note also that there is a placeholder function to write JSON data to
disk - this will later be replaced with an actual function residing in
in `tasks/io.py`