Return response after retrying even on failure
`retry` should only decorate the get function. It should not make the
decision to discard the response if the request fails after retries.
The decision on what to do with the failed response should be made
outside of the `retry` decorator, so we return the response if retrying
has been unsuccessful.

In our case, we want to discard failed responses and move on, so we
check the response status code after the retry and break the loop if
the status code is not 200.