Make the `get_workflow_runs` task a script
Add an `if __name__ == "__main__":` block to `tasks.get_workflow_runs`
so that it can be invoked as a script. Running the script extracts
workflow runs data from GitHub and writes them to `DATA_DIR`.

Since a token is needed to access the workflow runs data of private
repositories, add instructions to the README on how to get the
appropriate token and set it as an environment variable.