Simplify extraction step in `get_workflow_runs`
Rewrite `main` to use nested for-loops and "mostly pure"[*] functions
in the extraction step.
We will rework this code to replace the use of nested for-loops with
better generator expressions.

In order to do the above, I had to use the same timestamp to name all
directories in the same extraction, which is a trade-off that we will
examine when we rework the code.

[*]: The only side effect that remains is errors being printed to the
terminal if the API response is not a 200. We will also review this side
effect when we further rework this task.