Require a github token to fetch workflow runs
Before this change, if one runs the `get_workflow_runs` task without
a GitHub token, the task would run but only fetch workflow runs of
public repos. There were no warnings in place to inform the user
that private repos are not being fetched and may lead to confusion for
downstream analysis.

Here, we error instead if the token cannot be retrieved from the
environment. Users will have to obtain a token and always fetch both
public and private repos' workflow runs.

See discussion in #70.