Pass github org as arg to `main` and `extract`
We switched to using `responses` to mock `requests` in the previous
commit.
Accidental calls to the real API might be made (if `@responses.activate`
was forgotten), in which case we don't want to be fetching pages and
pages of data for the real `opensafely` org in a test.

To be safe, never involve `opensafely` in tests. To do that easily, have
`org` as an argument to `main`, `extract`, `get_repos` and
`get_repo_workflow_runs`.