Make filter function independent of state
Previously the function containing filtering logic in the
`get_workflow_runs` task was dependent on the filesystem.

As advised by the footnote in #78, refactor the code such that the
filtering function does not browse the filesystem, but only takes an
iterable of filepaths as input and outputs an iterable (in this case a
generator) of filepaths.