Ignore jobs with empty run commands in get_jobs
Closes #55.

As the issue suggested, use a standalone function after `extract` and
before `get_records` to filter out jobs with empty run commands.

This was done because all jobs with empty run commands (that were
created in 2025) were not real jobs, but only exist to send the user
some form of notification.