Use `updated_at` timestamp of data for filename
Indexing the data by the timestamp at which the ETL timestamp was run is
inefficient and unnecessary, as data that hasn't changed between
extractions is not identified. This leads to writing the same data to
disk over and over again.

Both repo metadata and workflow runs data retrieved from GitHub contain
an `updated_at` timestamp. If we identify data using this timestamp and
using it for naming files, we can potentially write new files only when
the data has been updated on GitHub.