Avoid unnecessary file writes in `get_workflow_runs`
The changes in this commit are purely for efficiency.

The `get_workflow_runs` task writes a JSON file for every for
every repo and workflow run retrieved from the GitHub API. Some of the
files written are identical to files already written in previous
runs of the task.
We can check if a given file already exists, if so, skip writing it.