Adopt GitHub flow alongside CI
We record our decision to adopt GitHub flow alongside CI and remove the
`open_pr` job, which automatically opened a pull request for any commits
pushed to a non-mainline branch.

Many developers will know GitHub flow is "*the* way everyone uses PRs".
Nevertheless, it isn't the *only* way. (The similarly named [Gitflow][]
is an alternative.) For this reason, it's useful to link to a document
that describes GitHub flow.

[Gitflow]: https://www.atlassian.com/git/tutorials/comparing-workflows/gitflow-workflow