Replace function with generator expression
b8845411 ignored jobs with empty run commands, as per #55. It did so by
adding a function, but adding a function isn't necessary: we could
instead use a generator expression with a condition. Doing so is much
less work! Admittedly, we don't test the enclosing function (`main`). We
should, but we didn't test the enclosing function before b8845411
either.