Move `open_pr` job into new workflow
We could equally have said "Move `check` and `test` jobs into new
workflow", but you get the idea. We now have an upstream "Check and
Test" workflow (`check_and_test.yaml`) and a downstream "Continuous
Integration" workflow (`ci.yaml`).

Why? Because we will add a new "Deploy" workflow that is also downstream
of the "Check and Test" workflow, but not the "Continuous Integration"
workflow. This workflow will be triggered on push to `main`.