Add task for getting OpenCodelists logins
Connect to a copy of the OpenCodelists database and write out a fact
table for user logins to OpenCodelists.

- SQLAlchemy extracts times as timezone-naive `datetime` objects from
SQLite, but the times are actually in UTC. Hence, append the UTC
timezone in the ETL task.

- Hashing is used to avoid extracting email addresses in plain text;
SHA-256 was chosen somewhat arbitrarily as a relatively (to SHA-1)
secure hashing algorithm. Hashing 1000 email addresses (each of 30
characters) takes around a millisecond on my laptop, so I do not expect
much impact on performance.