Test task modules have required attributes
Task modules have an implicit interface: They have a `main` function and
are designed to be run on the command-line. This adds a test to ensure
that any modules we add to `tasks.tasks` in the future implement this
interface.

For simplicity, we test that a module has an attribute called main; this
attribute may or may not be a function. We could use `callable` to test
whether the attribute is also callable (functions are callables), but
that seems like a step too far.