Test `<task>.main` doesn't accept parameters
Task modules have an implicit interface (14557ac). To add a task runner
(#134), it's convenient to extend this interface such that the `main`
function doesn't accept parameters. Let's add a test to ensure that any
modules we add to `tasks.tasks` in the future implement this interface.
Because an existing module doesn't yet implement it, we mark the test as
`xfail`.