Add a task runner
This adds a skeleton (i.e. basic structure) task runner: a mechanism for
listing all tasks and running individual tasks. The intention is to
remove the need to document tasks in `README.md`, which is easy to
forget (cd26a1a), and to provide a location for further enhancements.

The task runner's `run` subcommand does no more than running a task on
the command-line would do; it doesn't, for example, handle errors. For
this reason, and because there's very little logic that isn't concerned
with parsing arguments, `tasks/__main__.py` is excluded from coverage.

Closes #134