Spike: Add Google Authentication
Streamlit is planning to but hasn't added functionality to pass secrets
to `st.login` other than via a secrets.toml file.

There are a number of workarounds proposed by the community in
streamlit ticket 10543; this commit uses one that accesses a private
attribute of the streamlit secrets singleton. It has the advantage of:
 - not needing any third party packages to write out TOML files
 - not needing to run extra commands in dev/prod before running the app
 - keeping all the environment variables in one place (`.env` file for
   dev and Digital Ocean settings in prod)
over the other workaround, which is to use a `.tmpl.toml` file.

I've chosen to raise `KeyError`s if any of the required environment
variables are missing rather than letting Streamlit raise its own error
later on, since Streamlit will refer to the "secrets.toml" file which
doesn't exist in this setup, which might be confusing.