SQLite -> DuckDB
This switches SQLite with DuckDB. We point DuckDB to CSVs with URIs
rather than paths, because we want to point to the local filesystem in
development but to an object storage server in production (by setting
the `REPOSITORY_ROOT_URI` environment variable). We have to jump through
a couple of Pandas-shaped hoops to ensure that `get_logins_per_day`
interpolates counts of zero for days without logins, but it's important
to do so here, because then we don't have to remember to do so
everywhere else.