Add a task for fetching PR details from the GitHub API
This deliberately leaves some duplication between the new task and the
existing one for workflow runs, as it's not clear yet what the useful
abstractions might really be.

The two tasks both fetch repository lists, but the volume of API calls
for that is so much less than for the PRs and workflow runs that it
doesn't seem worth eliminating the duplicate work.