Add `REPOSITORY_ROOT_URI` to app spec
We add the envvar to the app spec so it's available to the Streamlit app
at run time *and* so it persists across builds. Although it's not a
"secret" as such, we don't want to hard-code the value in the app spec,
so the value is passed from GitHub to the app spec as a repository
secret. We set its type to `SECRET` for consistency with the repository
secret. (It would be better to use a repository variable, but it seems
that it's not possible to pass the value from GitHub to the app spec in
this way.)

For more information about how the value is passed, see the `cd.deploy`
GitHub Actions job (`.github/workflows/cd.yaml`) and the "[Deploy an app
(with a referenced secret)][1]" section of the
digitalocean/app_action/deploy README.

[1]: https://github.com/digitalocean/app_action/tree/v2.0.10?tab=readme-ov-file#deploy-an-app-with-a-referenced-secret