Don't encrypt `REPOSITORY_ROOT_URI`
There's no need to encrypt it; it's useful to see its value in the
DigitalOcean and GitHub UIs. We just don't want to hard-code it in the
app spec.

To achieve this, we make it a repository variable in GitHub and set
its type to `GENERAL` in the app spec.