Check health via Streamlit script health endpoint
App Platform performs a "readiness" health check as it deploys the app
to see if it is ready to receive incoming traffic.

By default the check attempts a TCP connection to the app's `http_port`,
which will succeed even if the Streamlit app raises an error.
We want the health check to fail when Streamlit raises an error,
so we configure the check to make an HTTP request to Streamlit's
script health check endpoint instead.

We do not adjust other health check settings and leave them as
DigitalOcean's defaults (the app's `http_port` value will by default
be used as the port for the health check).

If the check fails, the app rolls back to the last healthy deployment.