Configure health check for streamlit app script
App Platform performs a "readiness" health check as it deploys the app
to see if it is ready to receive incoming traffic.

By default, the check attempts a TCP connection to the app's http port,
which will succeed even if the Streamlit app raises an error.
We want the health check to fail when Streamlit raises an error,
so we configure the check to make an HTTP request to Streamlit's
script health check endpoint instead.

If the check fails, the app rolls back to the last healthy deployment.

On the parameter choices:
- At the time of writing, the deployment takes around 40-50 seconds, so
use an initial delay of around twice that time (90 seconds) to be safe.

- Making multiple attempts to verify the script health seems unnecessary
and will clutter the logs, so set thresholds to 1.

- To compensate for the strict failure threshold, we set a long timeout
of 60 seconds.