Enable the app's script health check endpoint
`server.scriptHealthCheckEnabled` is a hidden streamlit config that
enables the `/_stcore/script-health-check ` endpoint for use. If the
streamlit app script throws an error, this endpoint will return a 503
status code, which we can use for health checking in deployment.

Note that this setting is subject to change according to streamlit, so
we should be careful when upgrading streamlit versions to see if
anything about it has changed.