Remove `tasks.tasks.get_job_requests`
This task isn't part of the MVP. If we need it again, then we can
retrieve it from Git.

This task is the only caller of `INDEX_DATE`, `get_repo`, and
`load_project_definition`, so we remove these too. It's also the only
caller of `jobserver_engine` and `jobserver_metadata`. Removing these
will require more work, so we will do so in a follow-up commit and
simply placate coverage.