Remove `tests.jobserver.tables`
This database isn't part of the MVP. If we need it again, then we can
retrieve it from Git. By removing it, we remove an awful lot of
complexity from out test suite and CI GitHub Actions workflow.

Removing `tests/conftest.py` upset pytest's test discovery, so we add
`tests/app/__init__.py` to compensate. Git thinks this is a rename,
which isn't correct, really.