Use consistent labels and formatting
We're updating *Python* dependencies; a single space between each step;
and capitalise Slack, because its a proper noun.

I've also removed the comment that tells us the version string of the
SHA, because this seems like it could become outdated. (At the very
least, we rely on something or someone to update it, so why bother?)