Add `get_codelist_create_events_per_day`
We add the `get_codelist_create_events_per_day` method to the
repository. We don't add this method to the `AbstractRepository`,
however; if we did so, then we'd have to add this method to each fake
repository, and then ensure each fake repository was tested (for
coverage). That's a lot of unnecessary work and suggests that we're not
getting value for money from `AbstractRepository`.

`get_codelist_create_events_per_day` is very similar to
`get_logins_per_day`. However, it's not possible to parametrise the
underlying query because parameters can't be used as expressions.
Consider the following example:

```
CAST(created_at AS DATE)
```

In the above, `created_at` is an expression. Let's parametrise it:

```
CAST($col_name AS DATE)
```

When the above is evaluated, the result is:

```
CAST("created_at" AS DATE)
```

In the above, `"created_at"` is a string literal. You can find out more
about parameters in DuckDB's [DB API][].

The solution is, I think, to use DuckDB's [Expression API][]. However,
it's more important right now to get this working, than to get this
working nicely, so we'll refactor later.

[DB API]: https://duckdb.org/docs/stable/clients/python/dbapi
[Expression API]: https://duckdb.org/docs/stable/clients/python/expression