Break out timeseries into function
There's nothing fancy here: We simply wrap the code in a function, and
pass in the data frame as an argument. We don't even move the function
elsewhere! If we did, then we'd probably have to write a separate test
for it, and it's not clear how we should test an Altair chart.