Use `just run` to run both the app and tasks
Generalise the `just run` command so that it can call both the
`streamlit` and `python` executables depending on the arguments. This
centralises the setup for running things in one place (the `justfile`).

The API of `just run` is deliberately chosen to match that of `uv run`,
so that when we eventually switch to using `uv`, we can rewrite
`just run` to be a thin wrapper around `uv run`.

This leads to us needing longer commands to run tasks or the app, but
we can add aliases in a later commit to shorten them again.